refactor: 현재 유저 정보 커스텀 훅
