+++ src/components/header/Header.jsx
@@ -1,27 +1,13 @@
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import './Header.css';
 import Navbar from "./Navbar";
 import logo from '/logo.png';
-import { useUserStore } from '../../store/useUserStore';
-import { useEffect } from 'react';
-import { useCookies } from 'react-cookie';
-import { getMyInfo } from '../../apis/authApi';
+import useMyInfo from '../../hooks/useMyInfo';
+
 
 function Header() {
   const navigate = useNavigate();
-  const { info, setInfo } = useUserStore();
-  const [ cookies ] = useCookies(['accessToken', 'refreshToken']);
-  const location = useLocation();
-
-  useEffect(() => {
-    getMyInfo(cookies.accessToken)
-    .then((res) => {
-      setInfo(res.data);
-    })
-    .catch((err) => {
-      alert(err.message);
-    })
-  }, [cookies, location])
+  const { info } = useMyInfo();
 
   return (
     <header>
